backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.2.1
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
uritemplate==4.2.0
